+++ app/model/knock/auth_token.rb
@@ -18,11 +18,11 @@ module Knock
     end
 
     def entity_for entity_class
-      if entity_class.respond_to? :find_for_authentication
-        entity_class.find_for_authentication @payload
+      if entity_class.respond_to? :from_token_payload
+        entity_class.from_token_payload @payload
       else
         if entity_class.to_s == "User" && Knock.respond_to?(:current_user_from_token)
-          warn "[DEPRECATION]: `Knock.current_user_from_token` is deprecated. Please implement `User.find_for_authentication` instead."
+          warn "[DEPRECATION]: `Knock.current_user_from_token` is deprecated. Please implement `User.from_token_payload` instead."
           Knock.current_user_from_token.call @payload
         else
           entity_class.find @payload['sub']